[Bronze V] Title: 이상한 기호, Time: 172 ms, Memory: 17704 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,13 +4,13 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
 
-        int result = calc(a, b);
+        Long result = calc(a, b);
         System.out.println(result);
     }
-    public static int calc(int a, int b){
+    public static Long calc(long a, long b){
         return (a + b) * (a - b);
     }
 }
